+++ package.json
@@ -3,22 +3,28 @@
   "description": "LiveACID MongoDB Connection Module",
   "version": "1.0.4",
   "private": true,
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "test": "node index.js"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc"
   },
   "engines": {
-    "node": "16.13.2"
+    "node": "16.14.0"
   },
   "dependencies": {
     "connect-mongo": "^4.6.0",
-    "dotenv": "^16.0.1",
+    "dotenv": "^16.0.3",
     "express-session": "^1.17.3",
-    "mongoose": "^5.13.14"
+    "mongoose": "^6.8.1",
+    "typescript": "^4.9.4"
   },
-  "main": "index.js",
   "directories": {
     "test": "test"
   },
   "author": "",
-  "license": "ISC"
+  "license": "ISC",
+  "devDependencies": {
+    "@types/express-session": "^1.17.5"
+  }
 }
