+++ src/session.ts
@@ -1,18 +1,17 @@
-const MongoStore = require('connect-mongo');
-const expressSession = require('express-session');
-const mongodb = require('./mongodb');
-const config = require('./config');
+import MongoStore from 'connect-mongo';
+import expressSession from 'express-session'
+import * as mongodb from './mongodb'
+import config from './config'
 
 module.exports = {
     session() {
         return expressSession({
             secret: config.session.secret,
-            maxAge: config.session.maxAge,
             resave: config.session.resave,
             saveUninitialized: config.session.saveUninitialized,
             store: MongoStore.create({
-                collection: config.session.collection,
-                clientPromise: mongodb.client(),
+                collectionName: config.session.collection,
+                clientPromise: mongodb.client()
             }),
             cookie: {
                 maxAge: config.session.maxAge,
